Create FourSum.java
Added a solution for a problem called foursum in java

Approach 2: Hash Set
Intuition
Since elements must sum up to the exact target value, we can also use the One-pass Hash Table approach.
In 3Sum: Hash Set, we solved the problem without sorting the array. To do that, we needed to sort values within triplets, and track them in a hash set. Doing the same for k values could be impractical.
So, for this approach, we will also sort the array and skip duplicates the same way as in the Two Pointers approach above.  Since the array is sorted, we can just compare the found pair with the last one in the result res


Time Complexity: O(nk−1)O(n^{k - 1})O(n 
k−1
 ), or O(n3)O(n^3)O(n 
3
 ) for 4Sum. We have k−2k - 2k−2 loops iterating over nnn elements, and twoSum is O(n)O(n)O(n).

Note that for k>2k > 2k>2, sorting the array does not change the overall time complexity.

Space Complexity: O(n)O(n)O(n) for the hash set. The space needed for the recursion will not exceed O(n)O(n)O(n).